+++ lib/src/builder/elements/column/foreign_column_element.dart
@@ -50,4 +50,7 @@ class ForeignColumnElement extends ColumnElement
       return true;
     }
   }
+
+  @override
+  String? get defaultValue => null;
 }