+++ lib/src/cli/migration/output.dart
@@ -23,7 +23,7 @@ Future<void> outputSchema(Directory dir, DatabaseSchemaDiff diff) async {
       createTables +=
           ''
           'CREATE TABLE IF NOT EXISTS "${table.name}" (\n'
-          '${table.columns.values.map((c) => '  "${c.name}" ${c.type} ${c.isNullable ? 'NULL' : 'NOT NULL'}').join(",\n")}\n'
+          '${table.columns.values.map((c) => '  "${c.name}" ${c.type} ${c.isNullable ? 'NULL' : 'NOT NULL'}${c.defaultValue != null ? ' DEFAULT ${c.defaultValue}' : ''}').join(",\n")}\n'
           ');';
     }
 
@@ -60,14 +60,22 @@ Future<void> outputSchema(Directory dir, DatabaseSchemaDiff diff) async {
         }),
         ...table.columns.modified.expand((c) sync* {
           if (c.prev.type != 'serial' && c.newly.type == 'serial') {
-            yield 'ALTER COLUMN "${c.prev.name}" SET DATA TYPE int8 USING ${c.newly.name}::int8';
+            yield 'ALTER COLUMN "${c.prev.name}" SET DATA TYPE int4 USING ${c.newly.name}::int4';
             yield "ALTER COLUMN \"${c.prev.name}\" SET DEFAULT nextval('${table.name}_${c.newly.name}_seq')";
           } else {
-            var update =
-                c.prev.type != c.newly.type
-                    ? 'SET DATA TYPE ${c.newly.type} USING ${c.newly.name}::${c.newly.type}'
-                    : '${c.newly.isNullable ? 'DROP' : 'SET'} NOT NULL';
-            yield 'ALTER COLUMN "${c.prev.name}" $update';
+            if (c.prev.type != c.newly.type) {
+              yield 'ALTER COLUMN "${c.prev.name}" SET DATA TYPE ${c.newly.type} USING ${c.newly.name}::${c.newly.type}';
+            }
+            if (c.prev.isNullable != c.newly.isNullable) {
+              yield 'ALTER COLUMN "${c.prev.name}" ${c.newly.isNullable ? 'DROP' : 'SET'} NOT NULL';
+            }
+            if (c.prev.defaultValue != c.newly.defaultValue) {
+              if (c.newly.defaultValue != null) {
+                yield 'ALTER COLUMN "${c.prev.name}" SET DEFAULT ${c.newly.defaultValue}';
+              } else {
+                yield 'ALTER COLUMN "${c.prev.name}" DROP DEFAULT';
+              }
+            }
           }
         }),
       ];